svn import #54126

+++ engines/lastexpress/entities/pascale.cpp
@@ -19,7 +19,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  * $URL: https://scummvm.svn.sourceforge.net/svnroot/scummvm/scummvm/trunk/engines/lastexpress/entities/pascale.cpp $
- * $Id: pascale.cpp 53579 2010-10-18 19:17:38Z sev $
+ * $Id: pascale.cpp 53880 2010-10-27 19:19:22Z littleboy $
  *
  */
 
@@ -816,7 +816,7 @@ IMPLEMENT_FUNCTION(26, Pascale, chapter4Handler)
 label_callback1:
 			if (ENTITY_PARAM(1, 2) && ENTITY_PARAM(1, 4)) {
 				if (!params->param3)
-					params->param3 = getState()->time + 9000;
+					params->param3 = (uint)(getState()->time + 9000);
 
 				if (params->param5 != kTimeInvalid) {
 
@@ -828,7 +828,7 @@ label_callback1:
 					}
 
 					if (!getEntities()->isInRestaurant(kEntityPlayer) || !params->param5)
-						params->param5 = getState()->time;
+						params->param5 = (uint)getState()->time;
 
 					if (params->param5 < getState()->time) {
 						params->param5 = kTimeInvalid;
